Merging the latest features into dev

+++ app/Http/Controllers/Teacher/LessonController.php
@@ -63,12 +63,15 @@ class LessonController extends Controller
 
         // Retrieve the videos associated with the lesson
         $videos = $lesson->videos;
+        // Retrieve the documents associated with the lesson
+        $documents = $lesson->documents;
 
         // Pass the lesson details to the view
         return view('teacher.lessons.show', [
             'lesson' => $lesson,
             'course' => $lesson->course,
             'videos' => $videos,
+            'documents' => $documents,
             'previousLesson' => $previousLesson,
             'nextLesson' => $nextLesson
         ]);
@@ -111,7 +114,7 @@ class LessonController extends Controller
             'videos.*.url' => 'nullable|string|url',
             'videos.*.description' => 'nullable|string',
             'documents.*.title' => 'nullable|string|max:255',
-            'documents.*.file' => 'nullable|file|mimes:pdf,doc,docx,xls,xlsx|max:2048',
+            'documents.*.file' => 'nullable|file|mimes:pdf,doc,docx,xls,xlsx,txt|max:2048|required_with:documents.*.title',
             'documents.*.description' => 'nullable|string',
             'course_id' => 'required|exists:courses,id',
             'section_id' => 'nullable|exists:sections,id',
@@ -160,21 +163,21 @@ class LessonController extends Controller
                 }
             }
 
-            // /* Handle documents */
-            // if (isset($validated['documents']) && is_array($validated['documents'])) {
-            //     foreach ($validated['documents'] as $index => $document) {
-            //         if (isset($document['file']) && $request->hasFile("documents.$index.file")) {
-            //             $file = $request->file("documents.$index.file");
-            //             $path = $file->store('documents', 'public');
-            //             Document::create([
-            //                 'title' => $document['title'] ?? 'document for ' . $lesson->title,
-            //                 'file' => $path,
-            //                 'description' => $document['description'] ?? 'A document for the lesson titled "' . $lesson->title . '".',
-            //                 'lesson_id' => $lesson->id,
-            //             ]);
-            //         }
-            //     }
-            // }
+            /* Handle documents */
+            if (isset($validated['documents']) && is_array($validated['documents'])) {
+                foreach ($validated['documents'] as $index => $document) {
+                    if (isset($document['file']) && $request->hasFile("documents.$index.file")) {
+                        $file = $request->file("documents.$index.file");
+                        $path = $file->store('documents', 'public');
+                        Document::create([
+                            'title' => $document['title'] ?? 'Document for ' . $lesson->title,
+                            'file' => $path,
+                            'description' => $document['description'] ?? 'A document for the lesson titled "' . $lesson->title . '".',
+                            'lesson_id' => $lesson->id,
+                        ]);
+                    }
+                }
+            }
         } catch (\Exception $e) {
             // In case of creation error, redirection with error message
             return redirect()->route('teacher.lessons.index')
@@ -229,7 +232,7 @@ class LessonController extends Controller
             'video_url.*' => 'nullable|string|url',
             'video_description.*' => 'nullable|string',
             'document_title.*' => 'nullable|string|max:255',
-            'document_file.*' => 'nullable|file|mimes:pdf,doc,docx,txt|max:2048', // Example for allowed file types and size
+            'document_file.*' => 'nullable|file|mimes:pdf,doc,docx,xls,xlsx,txt|max:2048', // Example for allowed file types and size
             'document_description.*' => 'nullable|string',
             'course_id' => 'sometimes|required|exists:courses,id',
             'section_id' => 'nullable|exists:sections,id',
@@ -306,7 +309,7 @@ class LessonController extends Controller
                     if ($documentId && isset($existingDocuments[$documentId])) {
                         $document = $existingDocuments[$documentId];
                         // Delete the file from storage
-                        if (Storage::exists($document->file_path)) {
+                        if (!empty($document->file_path) && Storage::exists($document->file_path)) {
                             Storage::delete($document->file_path);
                         }
                         $document->delete();
@@ -337,7 +340,7 @@ class LessonController extends Controller
                 // Check if the document was marked for deletion
                 if (!in_array($document->id, array_column($documents, '_id'))) {
                     // Delete the file from storage
-                    if (Storage::exists($document->file_path)) {
+                    if (!empty($document->file_path) && Storage::exists($document->file_path)) {
                         Storage::delete($document->file_path);
                     }
                     $document->delete();
